Rename project to rimecmd
When thinking about the naming, it was thought that this project
is to be mainly used as a daemon in editors, thus the name rimed.
However, now the development is mostly done to make it work with
a terminal. Thus the renaming. cmd mean command, which can both
be using in a terminal and in a pipe as a daemon.