Implement backspacing in terminal interface
Though it is described as implementing backspacing, no special handling of
backspacing is added in this commit. This commit just adds partial support to
rime's composition, which is a description of the editing box of the input
method. More functions of composition, such as moving the cursor, still needs
to be supported.