Let pre-commit script use its own directory
pre-commit uses the RUSTFLAGS environment variable. On command line, generally
no or different flags are set. This causes the script to always have to
recompile all source code since what it compiled the last time is overwritten
by the developer's compilation, which are run under different flags. That takes
a lot of time.