Assign exit code 2 to the one input closed error
When rimecmd listens to both tty and stdin, it can't continue when
one of the input is closed, so when this happens, this is an error.
A user might want to catch this specific error, as they may want
to stop rimecmd by closing one of the input, and know that rimecmd
exited because of this instead of some other problems. Therefore,
a specific error code is assigned to this case.